requirements.txt
certifi==2021.10.8
click==8.1.2
dominate==2.6.0
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
requests==2.27.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.1
WTForms==3.0.1
gunicorn==20.1.0
setuptools==60.2.0
charset-normalizer==2.0.12
wheel==0.37.1
python-dotenv~=0.20.0